feat: clean up and cosolidate tests
